+++ lua/slimline/components/mode.lua
@@ -7,15 +7,15 @@ local config = require('slimline').config
 --- @param direction string
 --- |'"right"'
 --- |'"left"'
+--- @param hls table
 --- @return string
-function M.render(sep, direction)
+function M.render(sep, direction, hls)
   local mode = utils.get_mode()
   local render = mode
-  if config.verbose_mode == false then
+  if config.configs.mode.verbose == false then
     render = string.sub(mode, 1, 1)
   end
-  local hl = highlights.get_mode_hl(mode)
-  return highlights.hl_component({ primary = render }, hl, sep, direction)
+  return highlights.hl_component({ primary = render }, hls, sep, direction)
 end
 
 return M