+++ lua/slimline/components/path.lua
@@ -1,4 +1,6 @@
 local M = {}
+
+local name = 'path'
 local highlights = require('slimline.highlights')
 local config = require('slimline').config
 
@@ -6,25 +8,34 @@ local config = require('slimline').config
 --- @param direction string
 --- |'"right"'
 --- |'"left"'
+--- @param hls table
 --- @return string
-function M.render(sep, direction)
+function M.render(sep, direction, hls)
   if vim.bo.buftype ~= '' then
     return ''
   end
   local file = vim.fn.expand('%:t')
+
+  local icons = config.configs[name].icons
+
   if vim.bo.modified then
-    file = file .. ' ' .. config.icons.buffer.modified
+    file = file .. ' ' .. icons.modified
   end
   if vim.bo.readonly then
-    file = file .. ' ' .. config.icons.buffer.read_only
+    file = file .. ' ' .. icons.read_only
   end
-  local path = vim.fs.normalize(vim.fn.expand('%:.:h'))
-  if #path == 0 then
-    return ''
-  end
-  path = config.icons.folder .. path
 
-  return highlights.hl_component({ primary = file, secondary = path }, highlights.hls.component, sep, direction)
+  local path = nil
+
+  if config.configs.path.directory == true then
+    path = vim.fs.normalize(vim.fn.expand('%:.:h'))
+    if #path == 0 then
+      return ''
+    end
+    path = icons.folder .. path
+  end
+
+  return highlights.hl_component({ primary = file, secondary = path }, hls, sep, direction)
 end
 
 return M