+++ lua/slimline/init.lua
@@ -1,5 +1,8 @@
 local M = {}
 
+local highlights = require('slimline.highlights')
+local utils = require('slimline.utils')
+
 ---@alias Component function
 
 ---@type table<string, Component[]>
@@ -8,6 +11,25 @@ M.components = {
   center = {},
   right = {},
 }
+
+---@param component string
+---@return table
+local function get_sep(component)
+  local sep = {
+    left = M.config.sep.left,
+    right = M.config.sep.right,
+  }
+  local style = M.config.style
+
+  if M.config.configs[component] and M.config.configs[component].style ~= nil then
+    style = M.config.configs[component].style
+  end
+  if style == 'fg' then
+    sep.left = ''
+    sep.right = ''
+  end
+  return sep
+end
 
 ---@param component string | function
 ---@param position string?
@@ -23,24 +45,22 @@ local function get_component(component, position, direction)
   elseif type(component) == 'string' then
     local exist, component_module = pcall(require, string.format('slimline.components.%s', component))
     if exist then
-      local sep = {
-        left = M.config.sep.left,
-        right = M.config.sep.right,
-      }
-      if M.config.style == 'fg' then
-        if not ((component == 'mode' or component == 'progress') and not M.config.mode_follow_style) then
-          sep.left = ''
-          sep.right = ''
-        end
+      if M.config.configs[component].follow ~= nil then
+        component = M.config.configs[component].follow
       end
+      local sep = get_sep(component)
       if M.config.sep.hide.first and position == 'first' then
         sep.left = ''
       end
       if M.config.sep.hide.last and position == 'last' then
         sep.right = ''
       end
+      local hls = highlights.hls.components[component]
+      if component == 'mode' then
+        hls = highlights.get_mode_hl(utils.get_mode())
+      end
       return function(...)
-        return component_module.render(sep, direction, ...)
+        return component_module.render(sep, direction, hls, ...)
       end
     else
       return function()
@@ -104,6 +124,74 @@ local function get_components(components, group_position)
   return result
 end
 
+---Migrate options between versions and notify about deprecated options
+---@param opts table
+local function migrate_opts(opts)
+  local warnings = {}
+
+  if opts.verbose_mode ~= nil then
+    opts.configs.mode.verbose = opts.verbose_mode
+    table.insert(warnings, '`verbose_mode` deprecated. Use `configs.mode.verbose` instead.')
+  end
+
+  if opts.mode_follow_style ~= nil then
+    if opts.mode_follow_style == true then
+      opts.configs.mode.style = opts.style
+    else
+      opts.configs.mode.style = 'bg'
+    end
+    table.insert(warnings, '`mode_follow_style` deprecated. Use `configs.mode.style` instead.')
+  end
+
+  if opts.workspace_diagnostics ~= nil then
+    opts.configs.diagnostics.workspace = opts.workspace_diagnostics
+    table.insert(warnings, '`workspace_diagnostics` deprecated. Use `configs.diagnostics.workspace` instead.')
+  end
+
+  if opts.icons ~= nil then
+    table.insert(warnings, '`icons` deprecated. Use `configs.<component>.icon(s)` instead.')
+    if opts.icons.diagnostics then
+      opts.configs.diagnostics.icons =
+        vim.tbl_deep_extend('force', opts.configs.diagnostics.icons, opts.icons.diagnostics)
+    end
+    if opts.icons.git then
+      opts.configs.git.icons = opts.icons.git
+    end
+    if opts.icons.folder then
+      opts.configs.path.icons.folder = opts.icons.folder
+    end
+    if opts.icons.lines then
+      opts.configs.progress.icon = opts.icons.lines
+    end
+    if opts.icons.recording then
+      opts.configs.recording.icon = opts.icons.recording
+    end
+    if opts.icons.buffer then
+      if opts.icons.buffer.modified then
+        opts.configs.path.icons.modified = opts.icons.buffer.modified
+      end
+      if opts.icons.buffer.read_only then
+        opts.configs.path.icons.read_only = opts.icons.buffer.read_only
+      end
+    end
+  end
+
+  if opts.hl.modes then
+    table.insert(warnings, '`hl.modes` deprecated. Use `configs.mode.hl` instead.')
+    opts.configs.mode.hl = opts.hl.modes
+  end
+
+  if #warnings > 0 then
+    table.insert(
+      warnings,
+      '\nSee [here](https://github.com/sschleemilch/slimline.nvim/blob/main/lua/slimline/defaults.lua) for reference'
+    )
+    vim.notify(table.concat(warnings, '\n'), vim.log.levels.WARN, {
+      title = 'slimline.nvim',
+    })
+  end
+end
+
 ---@param opts table
 function M.setup(opts)
   if opts == nil then
@@ -114,11 +202,13 @@ function M.setup(opts)
     vim.o.showmode = false
   end
 
-  opts = vim.tbl_deep_extend('force', require('slimline.default_config'), opts)
+  opts = vim.tbl_deep_extend('force', require('slimline.defaults'), opts)
+
+  migrate_opts(opts)
 
   M.config = opts
 
-  require('slimline.highlights').create_hls()
+  highlights.create_hls()
 
   M.components.left = get_components(opts.components.left, 'left')
   M.components.center = get_components(opts.components.center, 'center')